+++ app/src/main/java/com/battlelancer/seriesguide/ui/episodes/EpisodeDetailsFragment.java
@@ -420,10 +420,11 @@ public class EpisodeDetailsFragment extends Fragment implements EpisodeActionsCo
 
         // dim text color for title if not released
         TextViewCompat.setTextAppearance(binding.textviewTitle, isReleased
-                ? R.style.TextAppearance_Title : R.style.TextAppearance_Title_Dim);
+                ? R.style.TextAppearance_SeriesGuide_Headline6
+                : R.style.TextAppearance_SeriesGuide_Headline6_Dim);
         if (!isReleased) {
             TextViewCompat.setTextAppearance(binding.textviewReleaseTime,
-                    R.style.TextAppearance_Caption_Dim);
+                    R.style.TextAppearance_SeriesGuide_Caption_Dim);
         }
 
         // guest stars
@@ -526,7 +527,7 @@ public class EpisodeDetailsFragment extends Fragment implements EpisodeActionsCo
         }
         bindingButtons.buttonEpisodeWatched.setOnClickListener(v -> onToggleWatched());
         bindingButtons.buttonEpisodeWatched
-                .setText(isWatched ? R.string.action_unwatched : R.string.action_watched);
+                .setText(isWatched ? R.string.state_watched : R.string.action_watched);
         CheatSheet.setup(bindingButtons.buttonEpisodeWatched, isWatched ? R.string.action_unwatched
                 : R.string.action_watched);
 
@@ -541,7 +542,7 @@ public class EpisodeDetailsFragment extends Fragment implements EpisodeActionsCo
         }
         bindingButtons.buttonEpisodeCollected.setOnClickListener(v -> onToggleCollected());
         bindingButtons.buttonEpisodeCollected.setText(collected
-                ? R.string.action_collection_remove : R.string.action_collection_add);
+                ? R.string.state_in_collection : R.string.action_collection_add);
         CheatSheet.setup(bindingButtons.buttonEpisodeCollected, collected
                 ? R.string.action_collection_remove : R.string.action_collection_add);
 
@@ -562,7 +563,7 @@ public class EpisodeDetailsFragment extends Fragment implements EpisodeActionsCo
             }
             bindingButtons.buttonEpisodeSkip.setOnClickListener(v -> onToggleSkipped());
             bindingButtons.buttonEpisodeSkip
-                    .setText(isSkipped ? R.string.action_dont_skip : R.string.action_skip);
+                    .setText(isSkipped ? R.string.state_skipped : R.string.action_skip);
             CheatSheet.setup(bindingButtons.buttonEpisodeSkip,
                     isSkipped ? R.string.action_dont_skip : R.string.action_skip);
         }