+++ app/src/main/java/com/battlelancer/seriesguide/ui/episodes/EpisodesAdapter.java
@@ -147,9 +147,11 @@ class EpisodesAdapter extends CursorAdapter {
 
         // dim text color if not released
         TextViewCompat.setTextAppearance(viewHolder.episodeTitle, isReleased
-                ? R.style.TextAppearance_Subhead : R.style.TextAppearance_Subhead_Dim);
+                ? R.style.TextAppearance_SeriesGuide_Subtitle1
+                : R.style.TextAppearance_SeriesGuide_Subtitle1_Dim);
         TextViewCompat.setTextAppearance(viewHolder.episodeAirdate, isReleased
-                ? R.style.TextAppearance_Body_Secondary : R.style.TextAppearance_Body_Dim);
+                ? R.style.TextAppearance_SeriesGuide_Body2_Secondary
+                : R.style.TextAppearance_SeriesGuide_Body2_Dim);
 
         // context menu
         viewHolder.contextMenu.setOnClickListener(v -> {