+++ app/src/main/java/com/battlelancer/seriesguide/ui/overview/OverviewFragment.java
@@ -62,6 +62,7 @@ import com.battlelancer.seriesguide.ui.episodes.EpisodeTools;
 import com.battlelancer.seriesguide.ui.episodes.EpisodesActivity;
 import com.battlelancer.seriesguide.ui.lists.ManageListsDialogFragment;
 import com.battlelancer.seriesguide.ui.search.SimilarShowsActivity;
+import com.battlelancer.seriesguide.ui.shows.RemoveShowDialogFragment;
 import com.battlelancer.seriesguide.ui.shows.ShowTools;
 import com.battlelancer.seriesguide.util.ClipboardTools;
 import com.battlelancer.seriesguide.util.DBUtils;
@@ -109,6 +110,7 @@ public class OverviewFragment extends Fragment implements
 
     @BindView(R.id.episode_empty_container) View containerEpisodeEmpty;
     @BindView(R.id.buttonOverviewSimilarShows) Button buttonSimilarShows;
+    @BindView(R.id.buttonOverviewRemoveShow) Button buttonRemoveShow;
 
     @BindView(R.id.episode_primary_container) View containerEpisodePrimary;
     @BindView(R.id.dividerHorizontalOverviewEpisodeMeta) View dividerEpisodeMeta;
@@ -206,6 +208,10 @@ public class OverviewFragment extends Fragment implements
         buttonSimilarShows.setOnClickListener(v ->
                 startActivity(SimilarShowsActivity.intent(requireContext(), showTvdbId, showTitle))
         );
+        buttonRemoveShow.setOnClickListener(v ->
+                RemoveShowDialogFragment
+                        .show(requireContext(), getParentFragmentManager(), showTvdbId)
+        );
 
         // episode buttons
         buttonWatchedUpTo.setVisibility(View.GONE); // Unused.
@@ -240,13 +246,9 @@ public class OverviewFragment extends Fragment implements
     public void onActivityCreated(Bundle savedInstanceState) {
         super.onActivityCreated(savedInstanceState);
 
-        // Are we in a multi-pane layout?
-        View seasonsFragment = requireActivity().findViewById(R.id.fragment_seasons);
-        boolean multiPane = seasonsFragment != null
-                && seasonsFragment.getVisibility() == View.VISIBLE;
-
-        // do not display show info header in multi pane layout
-        containerShow.setVisibility(multiPane ? View.GONE : View.VISIBLE);
+        // Hide show info if it is displayed in show fragment.
+        boolean isDisplayShowInfo = !getResources().getBoolean(R.bool.isFragmentShowNarrow);
+        containerShow.setVisibility(isDisplayShowInfo ? View.VISIBLE : View.GONE);
 
         LoaderManager loaderManager = LoaderManager.getInstance(this);
         loaderManager.initLoader(OverviewActivity.OVERVIEW_SHOW_LOADER_ID, null, this);
@@ -704,7 +706,7 @@ public class OverviewFragment extends Fragment implements
         } else {
             ViewTools.setVectorDrawableTop(buttonCollect, R.drawable.ic_collect_black_24dp);
         }
-        buttonCollect.setText(isCollected ? R.string.action_collection_remove
+        buttonCollect.setText(isCollected ? R.string.state_in_collection
                 : R.string.action_collection_add);
         CheatSheet.setup(buttonCollect, isCollected ? R.string.action_collection_remove
                 : R.string.action_collection_add);
@@ -870,7 +872,7 @@ public class OverviewFragment extends Fragment implements
         TvdbImageTools.loadShowPosterAlpha(requireContext(), imageBackground,
                 show.getString(ShowQuery.SHOW_POSTER_SMALL));
 
-        // regular network and time
+        // Regular network, release time and length.
         String network = show.getString(ShowQuery.SHOW_NETWORK);
         String time = null;
         int releaseTime = show.getInt(ShowQuery.SHOW_RELEASE_TIME);
@@ -887,8 +889,13 @@ public class OverviewFragment extends Fragment implements
             // "Mon 08:30"
             time = dayString + " " + timeString;
         }
+        String runtime = getString(
+                R.string.runtime_minutes,
+                String.valueOf(showCursor.getInt(ShowQuery.SHOW_RUNTIME))
+        );
+        String combinedString = TextTools.dotSeparate(TextTools.dotSeparate(network, time), runtime);
         TextView textViewNetworkAndTime = getView().findViewById(R.id.showmeta);
-        textViewNetworkAndTime.setText(TextTools.dotSeparate(network, time));
+        textViewNetworkAndTime.setText(combinedString);
         // set up long-press to copy text to clipboard (d-pad friendly vs text selection)
         ClipboardTools.copyTextToClipboardOnLongClick(textViewNetworkAndTime);