+++ app/src/main/java/com/battlelancer/seriesguide/ui/shows/ShowsFragment.java
@@ -34,11 +34,13 @@ import com.battlelancer.seriesguide.ui.movies.AutoGridLayoutManager;
 import com.battlelancer.seriesguide.ui.preferences.MoreOptionsActivity;
 import com.battlelancer.seriesguide.util.DBUtils;
 import com.battlelancer.seriesguide.util.ViewTools;
+import com.battlelancer.seriesguide.widgets.SgFastScroller;
 import com.google.android.material.floatingactionbutton.FloatingActionButton;
 import org.greenrobot.eventbus.EventBus;
 import org.greenrobot.eventbus.Subscribe;
 import org.greenrobot.eventbus.ThreadMode;
 import org.jetbrains.annotations.NotNull;
+import timber.log.Timber;
 
 /**
  * Displays the list of shows in a users local library with sorting and filtering abilities. The
@@ -67,6 +69,8 @@ public class ShowsFragment extends Fragment {
         View v = inflater.inflate(R.layout.fragment_shows, container, false);
 
         recyclerView = v.findViewById(R.id.recyclerViewShows);
+        new SgFastScroller(requireContext(), recyclerView);
+
         emptyView = v.findViewById(R.id.emptyViewShows);
         ViewTools.setVectorDrawableTop(emptyView, R.drawable.ic_add_white_24dp);
         emptyView.setOnClickListener(view -> startActivityAddShows());
@@ -131,8 +135,11 @@ public class ShowsFragment extends Fragment {
             boolean isEmpty = !adapter.getDisplayFirstRunHeader()
                     && (showItems == null || showItems.isEmpty());
             updateEmptyView(isEmpty);
+
+            // Once latest data loaded, schedule next refresh.
+            // Only while resumed (onResume/onPause also schedule/un-schedule refresh).
+            scheduleQueryUpdate(true);
         });
-        updateShowsQuery();
 
         // watch for sort order changes
         ShowsDistillationSettings.sortOrderLiveData
@@ -164,6 +171,7 @@ public class ShowsFragment extends Fragment {
     }
 
     private void updateShowsQuery() {
+        Timber.d("Running query update.");
         model.updateQuery(showFilter, ShowsDistillationSettings
                 .getSortQuery(showSortOrder.getSortOrderId(), showSortOrder.isSortFavoritesFirst(),
                         showSortOrder.isSortIgnoreArticles()));
@@ -196,22 +204,23 @@ public class ShowsFragment extends Fragment {
     public void onResume() {
         super.onResume();
 
-        // keep unwatched and upcoming shows from becoming stale
-        schedulePeriodicDataRefresh(true);
+        // Run query every time resuming to get latest data
+        // (e.g. displaying unwatched/upcoming shows depends on current time).
+        updateShowsQuery();
     }
 
     @Override
     public void onPause() {
         super.onPause();
 
-        // avoid CPU activity
-        schedulePeriodicDataRefresh(false);
     }
 
     @Override
     public void onStop() {
         super.onStop();
 
+        // avoid CPU activity
+        scheduleQueryUpdate(false);
         EventBus.getDefault().unregister(this);
     }
 
@@ -268,7 +277,7 @@ public class ShowsFragment extends Fragment {
             }
             case DISMISS: {
                 adapter.setDisplayFirstRunHeader(false);
-                model.reRunQuery();
+                updateShowsQuery();
                 break;
             }
         }
@@ -281,22 +290,19 @@ public class ShowsFragment extends Fragment {
      * underlying data, but because time has passed (e.g. relative time displays, release time has
      * passed).
      */
-    private void schedulePeriodicDataRefresh(boolean enableRefresh) {
+    private void scheduleQueryUpdate(boolean isPostNotRemove) {
+        Timber.d(isPostNotRemove ? "Scheduling query update." : "Removing planned query update.");
         if (handler == null) {
             handler = new Handler();
         }
         handler.removeCallbacks(dataRefreshRunnable);
-        if (enableRefresh) {
+        if (isPostNotRemove) {
             handler.postDelayed(dataRefreshRunnable, 5 * DateUtils.MINUTE_IN_MILLIS);
         }
     }
 
-    private Runnable dataRefreshRunnable = new Runnable() {
-        @Override
-        public void run() {
-            model.reRunQuery();
-        }
-    };
+    // Note: do not just re-run existing query, make sure timestamps in query are updated.
+    private Runnable dataRefreshRunnable = this::updateShowsQuery;
 
     private void startActivityAddShows() {
         startActivity(new Intent(getActivity(), SearchActivity.class).putExtra(