+++ app/src/main/java/com/battlelancer/seriesguide/widgets/FastScrollerDecoration.java
@@ -229,11 +229,6 @@ public class FastScrollerDecoration extends RecyclerView.ItemDecoration implemen
         return mState == STATE_VISIBLE;
     }
 
-    @VisibleForTesting boolean isHidden() {
-        return mState == STATE_HIDDEN;
-    }
-
-
     public void show() {
         switch (mAnimationState) {
             case ANIMATION_STATE_FADING_OUT:
@@ -247,10 +242,6 @@ public class FastScrollerDecoration extends RecyclerView.ItemDecoration implemen
                 mShowHideAnimator.start();
                 break;
         }
-    }
-
-    public void hide() {
-        hide(0);
     }
 
     @VisibleForTesting
@@ -283,6 +274,7 @@ public class FastScrollerDecoration extends RecyclerView.ItemDecoration implemen
                 || mRecyclerViewHeight != mRecyclerView.getHeight()) {
             mRecyclerViewWidth = mRecyclerView.getWidth();
             mRecyclerViewHeight = mRecyclerView.getHeight();
+
             // This is due to the different events ordering when keyboard is opened or
             // retracted vs rotate. Hence to avoid corner cases we just disable the
             // scroller when size changed, and wait until the scroll position is recomputed
@@ -315,7 +307,7 @@ public class FastScrollerDecoration extends RecyclerView.ItemDecoration implemen
             canvas.translate(mVerticalThumbWidth, top);
             canvas.scale(-1, 1);
             mVerticalThumbDrawable.draw(canvas);
-            canvas.scale(1, 1);
+            canvas.scale(-1, 1);
             canvas.translate(-mVerticalThumbWidth, -top);
         } else {
             canvas.translate(left, 0);
@@ -497,9 +489,15 @@ public class FastScrollerDecoration extends RecyclerView.ItemDecoration implemen
             return 0;
         }
         float percentage = ((newDragPos - oldDragPos) / (float) scrollbarLength);
-        int totalPossibleOffset = scrollRange - viewLength;
+        // FIXME: This is too small, likely missing bottom/end padding.
+        // Using scrollRange to compute scrollingBy also makes thumb track offset exactly.
+//        int totalPossibleOffset = scrollRange - viewLength;
+        //noinspection UnnecessaryLocalVariable
+        int totalPossibleOffset = scrollRange;
         int scrollingBy = (int) (percentage * totalPossibleOffset);
         int absoluteOffset = scrollOffset + scrollingBy;
+        // Note: it's fine to return a value even if scrolling is not possible,
+        // RecyclerView.scrollBy just won't scroll then.
         if (absoluteOffset < totalPossibleOffset && absoluteOffset >= 0) {
             return scrollingBy;
         } else {
@@ -512,7 +510,7 @@ public class FastScrollerDecoration extends RecyclerView.ItemDecoration implemen
         boolean insideOnYAxis = y <= mVerticalThumbHeight
                 || (y >= mVerticalThumbCenterY - mVerticalThumbHeight / 2
                 && y <= mVerticalThumbCenterY + mVerticalThumbHeight / 2);
-        return (isLayoutRTL() ? x <= mVerticalThumbWidth / 2
+        return (isLayoutRTL() ? x <= mVerticalThumbWidth
                 : x >= mRecyclerViewWidth - mVerticalThumbWidth)
                 && insideOnYAxis;
     }